Implemented snatching Multiple language and keep looking until a satisfied language is found
this is roughly working but probably needs loads of testing and fixing :)

Here's what it does:
When an episode is found it checks if it's an language we want and fetch it. If it's not a language specified as satisfactory the status of the episode is set to UNSATISFIED when it's processed to keep looking until a satisfactory one is found.